10/11/14_MR_14:26
Changed the "red" material to be a bit more striking and easier to see.
Added my test environment to the build settings (M.R_Test Env)

Added bounding boxes to keep players from bringing boxes with them.
Tweaked player killzones to fit the remaining part of the level.
Added a small section to the start of the level to house the initial control poster.
(Eg: Picture of the left thumbstick with "Use this to move" followed shortly by another thats a picture of the left trigger with "Use this to jump" etc.